Struct sass_rs::dispatcher::Dispatcher

pub struct Dispatcher {
    // some fields omitted
}

Holds the data structures needed to dispatch calls from libsass back in the Rust code.

Methods

impl Dispatcher

fn build(registry: Vec<(&'static str, Box<SassFunction>)>, sass_options: Arc<RwLock<SassOptions>>) -> Dispatcher

fn dispatch(&self) -> Result<(), String>

The dispatch function, this should be called until it returns an error. The caller should probably use a different thread.