Struct sass_rs::sass_value::SassValue

pub struct SassValue {
    // some fields omitted
}

Wrap a raw sass value.

Methods

impl SassValue

fn from_raw(raw: *const SassValueRaw) -> SassValue

fn sass_string(input: &str) -> SassValue

Create a raw SassValue containing a sass string.

fn sass_error(input: &str) -> SassValue

Create a raw SassValue containing a sass string.

fn as_raw(&mut self) -> Option<*mut SassValueRaw>

return a mutable raw, if available.

fn to_string(&self) -> Option<String>

Attempt to extract a String from the raw value.

fn to_vec_string(&self) -> Option<Vec<String>>

Attempt to extract a vector of strings from the raw value.

fn list_nth_to_string(&self, index: usize) -> Option<String>

Expect the SassValue to be a list and to contain a string, at the desired index.