Struct sass_rs::sass_context::SassOptions

pub struct SassOptions {
    pub raw: Unique<Struct_Sass_Options>,
}

Fields
raw: Unique<Struct_Sass_Options>
Methods

impl SassOptions

fn set_sass_functions(&mut self, sf: Vec<Sass_C_Function_Callback>)

Set the sass functions in the context, expects an array of tuples, each tuple contains the signature and function pointer.

fn set_output_style(&mut self, style: OutputStyle)

fn set_include_path(&mut self, path: PathBuf)